requirements.txt
openai
chromadb
python-dotenv
pillow
matplotlib
tqdm
numpy
fastapi
uvicorn
torch
torchvision
pytest
requests
pydantic
aiofiles
httpx
python-multipart
tabulate
pandas